Significant CMake refactor
Modularized build system so each cpp/ subdir
has its own CMake file, and added some CMake
options to control what subdirs are included
in the build.

The intent is to be able to conditionally build:
- the Python C/C++ bindings
- the C/C++ applications

We want the Python bindings for bundling in
a Pip wheel, whereas some of the C/C++ apps
rely on the site-specific MPI install and must
be built from source.