cibuildwheel updates - run-vcpkg action
Attempting to cache vcpkg install artifacts
to speed up the Github action build time.
A little bit tricky in the case of the
cibuildwheel linux target because it
builds in the context of a Docker container
hosted by the Github 'ubuntu-latest' runner.
For now, I'm hoping that the ubuntu-built
artifacts "just work" when linked to the
manylinux container build artifacts but we
will see.